+++ crates/oxc_linter/src/rules/import/export.rs
@@ -1,10 +1,11 @@
 use std::path::PathBuf;
 
+use rustc_hash::{FxHashMap, FxHashSet};
+
 use oxc_diagnostics::{LabeledSpan, OxcDiagnostic};
 use oxc_macros::declare_oxc_lint;
-use oxc_semantic::ModuleRecord;
 use oxc_span::{CompactStr, Span};
-use rustc_hash::{FxHashMap, FxHashSet};
+use oxc_syntax::module_record::ModuleRecord;
 
 use crate::{context::LintContext, rule::Rule};
