+++ crates/oxc_module_lexer/tests/integration/main.rs
@@ -71,7 +71,6 @@ pub fn parse(source: &str) -> ModuleLexer {
     let allocator = Allocator::default();
     let source_type = SourceType::mjs();
     let ret = Parser::new(&allocator, source, source_type).parse();
-    assert!(ret.errors.is_empty(), "{source} should not produce errors.\n{:?}", ret.errors);
     let module_lexer = oxc_module_lexer::ModuleLexer::new().build(&ret.program);
     // Copy data over because `ModuleLexer<'a>` can't be returned
     ModuleLexer {
