+++ crates/oxc_semantic/examples/semantic.rs
@@ -35,7 +35,6 @@ fn main() -> std::io::Result<()> {
     let program = parser_ret.program;
 
     let semantic = SemanticBuilder::new()
-        .build_module_record(path, &program)
         // Enable additional syntax checks not performed by the parser
         .with_check_syntax_error(true)
         .build(&program);
