+++ crates/oxc_semantic/src/checker/mod.rs
@@ -2,12 +2,13 @@ use oxc_ast::{
     ast::{DoWhileStatement, ForStatement, WhileStatement},
     AstKind,
 };
+use oxc_diagnostics::OxcDiagnostic;
+use oxc_span::Span;
 
 mod javascript;
 mod typescript;
 
 use javascript as js;
-pub use javascript::check_module_record;
 use typescript as ts;
 
 use crate::{builder::SemanticBuilder, AstNode};
@@ -117,5 +118,26 @@ pub fn check<'a>(node: &AstNode<'a>, ctx: &SemanticBuilder<'a>) {
             ts::check_ts_import_equals_declaration(decl, ctx);
         }
         _ => {}
+    }
+}
+
+#[cold]
+fn undefined_export(x0: &str, span1: Span) -> OxcDiagnostic {
+    OxcDiagnostic::error(format!("Export '{x0}' is not defined")).with_label(span1)
+}
+
+/// It is a Syntax Error if any element of the ExportedBindings of ModuleItemList
+/// does not also occur in either the VarDeclaredNames of ModuleItemList, or the LexicallyDeclaredNames of ModuleItemList.
+pub fn check_unresolved_exports(ctx: &SemanticBuilder<'_>) {
+    for reference_ids in ctx.unresolved_references.root().values() {
+        for reference_id in reference_ids {
+            let reference = ctx.symbols.get_reference(*reference_id);
+            let node = ctx.nodes.get_node(reference.node_id());
+            if node.flags().has_export_specifier() {
+                if let AstKind::IdentifierReference(ident) = node.kind() {
+                    ctx.errors.borrow_mut().push(undefined_export(&ident.name, ident.span));
+                }
+            }
+        }
     }
 }
