+++ crates/oxc_syntax/src/node.rs
@@ -82,6 +82,8 @@ bitflags! {
         const Class     = 1 << 1;
         /// Set functions containing yield statements
         const HasYield  = 1 << 2;
+        /// Set for `export { specifier }`
+        const ExportSpecifier  = 1 << 3;
     }
 }
 
@@ -102,5 +104,11 @@ impl NodeFlags {
     #[inline]
     pub fn has_yield(&self) -> bool {
         self.contains(Self::HasYield)
+    }
+
+    /// Returns `true` if this function has an export specifier.
+    #[inline]
+    pub fn has_export_specifier(&self) -> bool {
+        self.contains(Self::ExportSpecifier)
     }
 }
