+++ tasks/benchmark/benches/semantic.rs
@@ -1,5 +1,3 @@
-use std::path::Path;
-
 use oxc_allocator::Allocator;
 use oxc_benchmark::{black_box, criterion_group, criterion_main, BenchmarkId, Criterion};
 use oxc_parser::Parser;
@@ -22,10 +20,7 @@ fn bench_semantic(criterion: &mut Criterion) {
                     // We return `error`s to be dropped outside of the measured section, as usually
                     // code would have no errors. One of our benchmarks `cal.com.tsx` has a lot of errors,
                     // but that's atypical, so don't want to include it in benchmark time.
-                    let ret = SemanticBuilder::new()
-                        .with_build_jsdoc(true)
-                        .build_module_record(Path::new(""), &ret.program)
-                        .build(&ret.program);
+                    let ret = SemanticBuilder::new().with_build_jsdoc(true).build(&ret.program);
                     let ret = black_box(ret);
                     ret.errors
                 });
